Cart page design changes and removal of CartModal component
The cart page has been updated with a new design, including changes to the layout and styling of the cart items and the addition of a "You might also like" section. The CartModal component has been removed. The BlogPost component has also been updated with minor changes to the layout and styling of the "You might also like" section.