Remove closeModal function from CategoryModal props in ShopPage component
The diff shows the removal of the `closeModal` function from the props of the `CategoryModal` component in the `ShopPage` component. The `closeModal` function was previously used to close the modal when the search button was clicked in the `CategoryModal` component. However, it is no longer being used and has been removed to simplify the props of the `CategoryModal` component.

The `CategoryModal` component still receives the `active`, `categoryFilter`, `search`, and `setSearch` props, which are used to manage the state and behavior of the modal. The `setSearch` function is used to toggle the search state between true and false, while the `categoryFilter` function is used to filter the products based on the selected category. The `active` prop is used to keep track of the currently selected category.