"Add Helpers.CalculateTotal function and use it to calculate cart total; remove cartTotalPrice from store"
This diff adds a new function Helpers.CalculateTotal which calculates the total price of all items in the cart. It is used in the CartModal and Cart\_Section components to display the cart total. The cartTotalPrice field has been removed from the store and the total is now calculated on the fly using Helpers.CalculateTotal. The updateTotal function has been removed from the store and the updateTotal calls have been replaced with Helpers.CalculateTotal. The emptyCart function in the store has also been updated to not update the cartTotalPrice field.