"Update product image size and grid layout in ProductsPage and ShopPage"
This commit includes changes to the way product images are displayed in the ProductsPage and ShopPage components. In ProductsPage, the height of the main product image is now set to "h-full" instead of "h-auto". In ShopPage, the grid layout for product cards has been updated to include different gap sizes for different screen sizes. Additionally, the "justify-center" class has been added to the flex container for product cards on smaller screens to center them horizontally.